+++ src/commands/handlers/create-room.ts
@@ -1,6 +1,6 @@
 import { MessageType } from '../../socket-message/enums/message-type.enum';
-import { Users } from '../../api/users.api.service';
-import { Lobbies } from '../../api/lobbies.api.service';
+import { Users } from '../../api/users';
+import { Lobbies } from '../../api/lobbies';
 import { WSData } from '../interfaces/ws-data.interface';
 import { sendResponse } from '../send-response';
