+++ src/commands/handlers/reg.ts
@@ -1,7 +1,7 @@
 import { MessageType } from '../../socket-message/enums/message-type.enum';
-import { Users } from '../../api/users.api.service';
+import { Users } from '../../api/users';
 import { validateRegData } from '../validators/reg.validator';
-import { Lobbies } from '../../api/lobbies.api.service';
+import { Lobbies } from '../../api/lobbies';
 import { WSData } from '../interfaces/ws-data.interface';
 import { sendResponse } from '../send-response';
